0525 동기화

+++ src/page/Cs/NoticeModify.vue
@@ -27,7 +27,7 @@
             <p>{{ fileInfo }}</p>
             <p></p>
           </div>
-          <span><input type="checkbox" v-model="fileDel"> {{ FileNameSplit }}</span>
+          <span v-if="this.file"><p class="inline-block text-lg">파일제거</p><input type="checkbox" v-model="fileDel" class="mx-1.5">{{ FileNameSplit }}</span>
         </div>
         <div class="w-full flex justify-end px-6">
           <button class="bg-red-400 hover:bg-red-600 focus:ring-red-400 py-2 px-4 text-white font-semibold rounded-md shadow-md focus:outline-none focus:ring-2 focus:ring-opacity-75 text-xs sm:text-sm" @click="Modify">완료</button>
@@ -50,6 +50,8 @@ export default {
       file: "",
       isChecked : false,
       fixed : false,
+      fileDel: false,
+      fileInfo: ''
     }
   },
   methods: {
@@ -69,16 +71,38 @@ export default {
       }else{
         this.fixed = false
       }
-      if(this.fileDel){
-        storage.ref().child(`images/${this.FileNameSplit}`).delete()
-        this.file = ""
+      // if(this.fileDel){
+      //   storage.ref().child(`images/${this.FileNameSplit}`).delete()
+      //   this.file = ""
+      // }
+      // db.collection("notice").doc(this.$store.state.noticeId).update({
+      //   "name": this.name,
+      //   "title": this.title,
+      //   "content": this.content,
+      //   "date": this.date,
+      //   "fixed": this.fixed,
+      //   "file": this.file,
+      // }).then(() => {
+      //   alert("수정이 완료되었습니다.")
+      //   this.$router.replace("/cs/notice")
+      // })
+      if(this.file){
+        if(this.fileDel){
+          storage.ref().child(`images/${this.FileNameSplit}`).delete()
+          this.file = ""
+        }
+      }
+      // else if(this.file){
+      //   storage.ref().child(`images/${this.FileNameSplit}`).push()
+      // }
+      if((this.file != '')){
+        storage.ref().child(`images/${this.FileNameSplit}`).push()
       }
       db.collection("notice").doc(this.$store.state.noticeId).update({
         "name": this.name,
         "title": this.title,
         "content": this.content,
         "date": this.date,
-        "fixed": this.fixed,
         "file": this.file,
       }).then(() => {
         alert("수정이 완료되었습니다.")
@@ -100,10 +124,32 @@ export default {
     })
   },
   computed: {
+    // FileNameSplit(){
+    //   let parts;
+    //   let fileName;
+
+    //   if(this.file){
+    //     parts = this.file.split("%2F");
+    //     fileName = parts[parts.length - 1].split('?')[0]
+    //     return fileName
+    //   } else {
+    //     return ''
+    //   }
+    // }
     FileNameSplit(){
-      const parts = this.file.split("%2F");
-      const fileName = parts[parts.length - 1].split('?')[0]
-      return fileName
+      // const parts = this.file.split("%2F");
+      // const fileName = parts[parts.length - 1].split('?')[0]
+      // return fileName
+      let parts;
+      let fileName;
+
+      if(this.file){
+        parts = this.file.split("%2F");
+        fileName = parts[parts.length - 1].split('?')[0]
+        return fileName
+      } else {
+        return ''
+      }
     }
   },
 }
